requirements.txt
# Requirements for Telephony Mastery Content Processing System

# Core dependencies
pyyaml>=6.0
anthropic>=0.7.0

# Optional dependencies for enhanced functionality
markdown>=3.4.0
python-frontmatter>=1.0.0